requirements.txt
appdirs==1.4.4
asgiref==3.2.10
astroid==2.4.2
autopep8==1.5.4
bcrypt==3.2.0
beautifulsoup4==4.9.1
bidict==0.19.0
black==20.8b1
boto3==1.16.59
botocore==1.19.59
bs4==0.0.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
colorama==0.4.3
comtypes==1.1.7
confusable-homoglyphs==3.2.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.1.1
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.1
Django==3.1.1
django-allauth==0.42.0
django-crispy-forms==1.9.0
django-filter==2.3.0
django-model-utils==4.0.0
django-registration==3.1
django-rest-auth==0.9.5
django-rest-framework==0.1.0
django-rest-knox==4.1.0
djangorestframework==3.11.1
drf-yasg==1.20.0
dynaconf==3.1.2
filelock==3.0.12
future==0.18.2
geocoder==1.38.1
hvac==0.10.6
idna==2.10
inflection==0.5.1
isort==4.3.21
itypes==1.2.0
Jinja2==2.11.2
jmespath==0.10.0
jsonfield==3.1.0
knox==0.1.11
lazy-object-proxy==1.4.3
loc==0.0.2
loguru==0.5.3
Markdown==3.2.2
MarkupSafe==1.1.1
mccabe==0.6.1
mypy-extensions==0.4.3
numpy==1.20.1
oauthlib==3.1.0
packaging==20.4
pandas==1.2.3
paramiko==2.7.2
pathspec==0.8.1
pbr==5.5.1
pillow==8.1.0
pycodestyle==2.6.0
pycparser==2.20
PyNaCl==1.4.0
pyparsing==2.4.7
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.1
ratelim==0.1.6
regex==2020.11.13
requests==2.24.0
requests-oauthlib==1.3.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
s3transfer==0.3.4
shortuuid==1.0.1
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
stevedore==3.3.0
toml==0.10.1
typed-ast==1.4.2
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.25.9
validators==0.18.2
virtualenv==20.4.2
virtualenv-clone==0.5.4
virtualenvwrapper==4.8.4
win32-setctime==1.0.3
wrapt==1.12.1
